Optimize api server
